Tidying up code£
